Refactor ServerUtils inputs to be generic (SCRD-3292)
