+++ src/components/ServerUtils.js
@@ -18,8 +18,8 @@ import '../Deployer.css';
 import { translate } from '../localization/localize.js';
 import ServerTable from './ServerTable.js';
 import { isRoleAssignmentValid } from '../utils/ModelUtils.js';
-
-const TRUTHY = [true, 'true', 'True', 'yes', 'Yes'];
+import { LabeledDropdown } from './LabeledDropdown.js';
+import { ValidatingInput } from './ValidatingInput.js';
 
 export class SearchBar extends Component {
   constructor(props) {
@@ -140,267 +140,6 @@ export class ServerRolesAccordion extends Component {
   render() {
     return (
       <div className='roles-accordion'>{this.renderSections()}</div>
-    );
-  }
-}
-
-export class ServerInput extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      errorMsg: '',
-      inputValue: this.props.inputValue,
-      showMask: true
-    };
-  }
-
-  componentDidMount() {
-    if(this.props.updateFormValidity) {
-      let isValid = this.validateInput(this.state.inputValue, this.props);
-      //callback function from parent to initially check
-      //all inputs
-      this.props.updateFormValidity(this.props, isValid);
-    }
-  }
-
-  componentWillReceiveProps(newProps) {
-    if (this.props.inputValue !== newProps.inputValue) {
-      this.setState({inputValue : newProps.inputValue});
-      this.validateInput(newProps.inputValue, newProps);
-    }
-  }
-
-  isRequired = () => TRUTHY.includes(this.props.isRequired) || TRUTHY.includes(this.props.required)
-
-  validateInput(val, extraProps) {
-    let retValid = false;
-
-    if(this.isRequired()) {
-      if(val === undefined || val.length === 0) {
-        this.setState({errorMsg: translate('server.input.required.error')});
-        return retValid;
-      }
-    }
-
-    if(this.props.inputValidate && val !== '' && val !== undefined) {//have a validator and have some values
-      let validateObject = this.props.inputValidate(val, extraProps);
-      if (validateObject) {
-        if(validateObject.isValid) {
-          this.setState({errorMsg: ''});
-          retValid = true;
-        }
-        else {
-          this.setState({
-            errorMsg: validateObject.errorMsg
-          });
-        }
-      }
-      else {
-        this.setState({errorMsg: translate('server.validator.error')});
-      }
-    }
-    else {  //don't have validator
-      retValid = true;
-      this.setState({ errorMsg: ''});
-    }
-
-    return retValid;
-  }
-
-  handleInputChange = (e, props) => {
-    let val = e.target.value;
-    let valid = this.validateInput(val, props);
-    this.setState({
-      inputValue: val
-    });
-
-    //call back function from parent to handle the
-    //change...also it will call updateFormValidity
-    //to check all the inputs
-    this.props.inputAction(e, valid, props);
-  }
-
-  toggleShowHidePassword(e) {
-    let passwordField = e.target.previousSibling;
-    passwordField.type = this.state.showMask ? 'text' : 'password';
-    this.setState((prevState) => {return {showMask: !prevState.showMask};});
-  }
-
-  render() {
-    let inputType = 'text';
-    if(this.props.inputType) {
-      inputType = this.props.inputType;
-    }
-    let myprops = {};
-    if (inputType === 'number') {
-      myprops.min = this.props.min;
-      myprops.max = this.props.max;
-    }
-
-    let placeholder = this.props.placeholder;
-
-    let togglePassword = '';
-    if (inputType === 'password') {
-      if (this.state.showMask) {
-        togglePassword = <i className='material-icons password-icon'
-          onClick={(e) => this.toggleShowHidePassword(e)}>visibility</i>;
-      } else {
-        togglePassword = <i className='material-icons password-icon'
-          onClick={(e) => this.toggleShowHidePassword(e)}>visibility_off</i>;
-      }
-    }
-
-    if (TRUTHY.includes(this.props.disabled)) {
-      myprops.disabled = true;
-    }
-
-    if (TRUTHY.includes(this.props.autoFocus)) {
-      myprops.autoFocus = true;
-    }
-
-    if (this.isRequired()) {
-      myprops.required = true;
-
-      // append a '*' to the placeholder string
-      if (typeof(placeholder) === 'string' && !placeholder.endsWith('*')) {
-        placeholder += '*';
-      }
-    }
-
-    let classname = 'server-input';
-    if (this.props.moreClass) {
-      classname += ' ' + this.props.moreClass;
-    }
-
-    if (this.props.inputType === 'textarea') {
-      return (
-        <div className={classname}>
-          <textarea
-            id={this.props.id}
-            className='rounded-corner'
-            name={this.props.inputName}
-            value={this.state.inputValue}
-            onChange={(e) => this.handleInputChange(e, this.props)}
-            placeholder={placeholder}
-            disabled={this.props.disabled}
-            autoFocus={this.props.autoFocus}
-            {...myprops}>
-          </textarea>
-          <div className='error-message'>{this.state.errorMsg}</div>
-        </div>
-      );
-    } else {
-      return (
-        <div className={classname}>
-          <input
-            id={this.props.id}
-            className='rounded-corner'
-            type={inputType}
-            name={this.props.inputName}
-            value={this.state.inputValue}
-            onChange={(e) => this.handleInputChange(e, this.props)}
-            placeholder={placeholder}
-            disabled={this.props.disabled}
-            autoFocus={this.props.autoFocus}
-            {...myprops}>
-          </input>
-          {togglePassword}
-          <div className='error-message'>{this.state.errorMsg}</div>
-        </div>
-      );
-    }
-  }
-}
-
-export class ServerInputLine extends Component {
-  render() {
-    let labelStr = this.props.label ? translate(this.props.label) : '';
-    let label = this.props.label ? ((this.props.isRequired) ? labelStr + '*' : labelStr) : '';
-    return (
-      <div className='detail-line'>
-        <div className='detail-heading'>{label}</div>
-        <div className='input-body'>
-          <ServerInput isRequired={this.props.isRequired} inputName={this.props.inputName}
-            inputType={this.props.inputType} inputValidate={this.props.inputValidate} {... this.props}
-            inputAction={this.props.inputAction} inputValue={this.props.inputValue}
-            updateFormValidity={this.props.updateFormValidity} category={this.props.category}
-            moreClass={this.props.moreClass}/>
-        </div>
-      </div>
-    );
-  }
-}
-
-export class ServerDropdown extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      value: this.props.value
-    };
-  }
-
-  handleSelect = (e) => {
-    this.setState({value: e.target.value});
-    this.props.selectAction(e.target.value);
-  }
-
-  componentWillReceiveProps(newProps) {
-    if (this.state.value !== newProps.value) {
-      this.setState({value : newProps.value});
-    }
-  }
-
-  renderOptions() {
-    let options = this.props.optionList.map((opt) => {
-      if (this.props.defaultOption && opt === this.props.defaultOption.value) {
-        return <option key={opt} value={opt}>{this.props.defaultOption.label}</option>;
-      } else {
-        return <option key={opt} value={opt}>{opt}</option>;
-      }
-    });
-
-    if(this.props.emptyOption && (this.state.value === '' || this.state.value === undefined)) {
-      let emptyOption = [
-        <option
-          key='noopt' value={this.props.emptyOption.value}>{this.props.emptyOption.label}
-        </option>
-      ];
-      //add at the beginning
-      options = emptyOption.concat(options);
-    }
-
-    return options;
-  }
-
-  render() {
-    let classname = 'server-detail-select';
-    if (this.props.moreClass) {
-      classname += ' ' + this.props.moreClass;
-    }
-    return (
-      <div className={classname}>
-        <select className='rounded-corner' value={this.state.value} name={this.props.name}
-          onChange={this.handleSelect}>{this.renderOptions()}</select>
-      </div>
-    );
-  }
-}
-
-export class ServerDropdownLine extends Component {
-  render() {
-    let labelStr = this.props.label ? translate(this.props.label) : '';
-    let label = this.props.label ? ((this.props.isRequired) ? labelStr + '*' : labelStr) : '';
-    return (
-      <div className='detail-line'>
-        <div className='detail-heading'>{label}</div>
-        <div className='input-body'>
-          <ServerDropdown name={this.props.name} value={this.props.value}
-            optionList={this.props.optionList} emptyOption={this.props.emptyOption}
-            selectAction={this.props.selectAction} defaultOption={this.props.defaultOption}
-            moreClass={this.props.moreClass}/>
-        </div>
-      </div>
     );
   }
 }
