+++ src/styles/components/serverutils.less
@@ -142,18 +142,18 @@
   margin: 0 1em 2em 1em;
 }
 
-.server-input {
+.validating-input {
   .error-message {
     color:@errorcolor;
     display: block;
   }
 }
 
-.server-input.has-button input {
+.validating-input.has-button input {
   width: 70%;
 }
 
-.server-input input {
+.validating-input input {
   padding: 4px;
   width: 80%;
 }
