+++ src/styles/components/ymlvalidator.less
@@ -70,7 +70,7 @@
     }
   }
 
-  .server-input > input {
+  .validating-input > input {
     width: 90%;
   }
   .server-detail-select select {
